(Mtc) Change QoS Precondition Handling Policy for WLAN
The current policy in MtcPreconditionManager for WLAN handling is as
follows:
1. WLAN start or L2W HO
- WAIT_AUDIO_DEDICATED_BEARER timer is started.
- QosInfo is updated to AVAILABLE.
- QosStatusRecord (precondition attribute in SDP) is updated to sendrecv.
2. W2L HO
- WAIT_AUDIO_DEDICATED_BEARER timer is maintained.
- QosInfo is maintained.
- QosStatusRecord is maintained.

Problem: Due to this policy, if a W2L HO occurs exactly when the
existing timer expires, the call will be dropped without waiting for the
Dedicated EPS Bearer allocation complete.

Solution: The policy is changed as follows to improve reliability:
1. WLAN start or L2W HO
- WAIT_AUDIO_DEDICATED_BEARER timer is stopped.
- QosInfo is unchanged.
- QosStatusRecord is updated to sendrecv.
2. W2L HO
- WAIT_AUDIO_DEDICATED_BEARER timer is started.
- QosInfo is maintained.
- QosStatusRecord is maintained.

This change preserves the existing policy of always waiting for the
Dedicated EPS Bearer in Mobile RAT while simultaneously achieving a
lower probability of call drops.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulation W2L and L2W HO
Bug: 443701319
Flag: NONE pre-production
Change-Id: If109f6ae0a7a834fca56005ddfc14d7cfdacd38e