(Aos) Provides IMS supported networks with List<Integer> instead of int[]
This change refactors the way IMS supported network types are handled within DcSettings.
Previously, the list of supported network types was read as an integer array (int[]) and directly exposed in that format.
This modification changes the data structure used for external provision from an array (int[]) to a more idiomatic and flexible list (List<Integer>). This improves API consistency and ease of use for consumers of this setting by utilizing a standard Java collection type.

Flag: NONE pre-production
Bug: 447200032
Test: atest ImsStackJavaTests
Change-Id: I8f99659e8b51c2f6f540fdec06ec4f0396e61b7a