(core) Prefer UPDATE method for session refresh in a certain case
When the session refresh method is set to UPDATE preferred and UPDATE
method support is verified during an initial session setup, if the
mid-dialog INVITE transaction detects the absence of the Allow header,
the UPDATE method will not be used as the session refresh method. This
will result in a dropped call on the certain carrier network.

This change ensures that if UPDATE method support has been verified even
once in the same session, it will be maintained and used as the session
refresh method even if Allow header is missing.

Bug: 450828150
Flag: EXEMPT BUGFIX
Test: atest ImsStackNativeTests
Test: manual (verified UPDATE method is used for session refresh even if
incoming re-INVITE request doesn't contain Allow header)

Change-Id: I576017eb66eaa7049a25a4f3cc6c567119cad1dc