(mtc) Correct media direction for incoming call
Java MTC has a logic to convert remote direction to local when accepting
calls. But currently native MTC always notifies the Java with the local
direction, so this logic causes problems when incoming INVITE has
a=sendonly.

This CL removes this conversion, and also removes unused code to
restrict the usage of the modified function.

Test: atest ImsStackJavaTests
Test: Related scenarios using TISS (with/without 100rel)
      - Accept voice call (a=sendrecv/sendonly)
      - Accept video call (a=sendrecv/sendonly/recvonly)
        - Works as expected after disable
	  FEATURE_OVERRIDE_VIDEO_DIRECTION_FROM_CALL_TYPE.
          This behavior is same regardless of this CL.
Bug: 450796680
Flag: EXEMPT BUGFIX
Change-Id: Ib27c3c8eb058ba589751fac872a8357559464083