(Mtc) check remote precondition attributes for each media
For each media in the remote SDP, check if the precondition attributes are included,
If the precondition attribute is included, precondition attribute for the corresponding media will be included in the local SDP, and waiting remote QOS reserved for that media.
If the precondition attribute is not included, precondition attribute for the corresponding media will NOE be included in the local SDP, and NOT waiting remote QOS reserved for that media.

Test: verified MO, MT with or without certain media's precondition attributes using TISS
Test: atest ImsStackNativeEnablerMtcTests
Bug: 447276483
Flag: NONE pre-production

Change-Id: I72772209caa6dd5de06af10c5a2a3d6def88ff1b