(Mtc) Introcude ByeTransactionHandler for deferred operations
This change implements and adds the new ByeTransactionHandler class.
This handler is designed to execute a specific, pre-configured
operation upon the completion of a SIP BYE transaction.

- The handler is created and managed by the `IMtcCallController`.

- It is instantiated with the necessary operation injected at creation time.
Client classes specify the desired operation along with the conditions required
from the `ISession`.

- The handler performs the stored operation when `OnSessionInterfaceReleased()`
is called by the `SessionInterfaceHolder`, signifying the end of the session context.

Test: atest ImsStackNativeEnablerMtcTests
Test: normal call termination, BYE transaction timeout
Bug: 452167947
Flag: NONE pre-production
Change-Id: I6f6433df68dd511d7e63ac6ca2486336fc8f27ce