(Ssc) Adding a KEY_UT_USE_ROOT_ELEMENT_WHEN_DOC_QUERY_BOOL
This CL is to add KEY_UT_USE_ROOT_ELEMENT_WHEN_DOC_QUERY_BOOL
configuration. If this value is true, UE will add a node selector for a
root element of the XML document in the URI when UE sends HTTP GET
request to fetch an entire document.

Bug: 452260584
Test: atest SscUrlTest
Test: Manual test on live network
Flag: NONE pre-production
Change-Id: I863f52a2eea04d1cfc2cb2121b754996f86a55d7