(SmsStack) Fix MT SMS blocking after TR2 timer expiry
When the TR2 timer expires in `SmsRLStateMachine`, the native
`MtsMessageController` is not notified, leaving the `m_bProcessingMsg`
flag set to true. This blocks the processing of any new incoming MT SMS
messages, which are queued indefinitely.

This CL introduces a new JNI call from the Java layer to the native
layer to clear the stale message and reset the processing flag upon TR2
timer expiration.

- Added `clearStaleMessages()` in `MtsController.java` which is called from `SmsRLStateMachine.java` when the TR2 timer expires.
- This triggers a JNI event to call the new `ClearStaleMtSmsAndProcessNext()` method in `MtsMessageController.cpp`.
- `ClearStaleMtSmsAndProcessNext()` removes the stale MT SMS from the queue, resets `m_bProcessingMsg` to false, and triggers the processing of the next message.

Bug: 443833611
FLAG: NONE pre-production
Doc: go/notify-tr2-expiry-mts
Test: atest ImsStackNativeTests, ImsStackJavaTests
Test: Manually tested by preventing RP-ACK from being sent

Change-Id: I05e120ecb0b2bd8a76d9add2208bc7c8f0c4abb2