(Aos) Ignore MTU from Network for THREE
The UE is failing IMS registration because the network
assigned an MTU of 1900, causing the UE to correctly
use UDP for its SIP REGISTER message, but the network
unexpectedly requires TCP and thus provides no response.

Add a new config ignore_mtu_from_network_bool to ignore
MTU from Network for carrier Three.
If the value is true, UE will ignore the MTU from network.
The default is false and the value for Three is true.

Bug: 454900261
Test: b/454900261#comment9 and atest ImsStackNativeEnablerAosTests
Flag: NONE pre-production
Change-Id: I5156d2e9997ef6c108871295c48347943310bede