(Mtc) Set Session as terminated if another session is established
This commit changes to invoke
IMtcSession#SetSessionTerminatedOrStartFailed when another session is
established. In this case, this 'forked' or 'original' session is
immediately removed without any further SIP signalling. So, the
SessionInterfaceHolder doesn't need to handle this with guard timer.

Issue: SessionInterfaceHolder waits the guard timer with Timer F period
so stopImsTraffic cannot be called within this duration.

Test: atest ImsStackNativeEnablerMtcTests
Test: forking scenario, normal single call scenario
Bug: 453538063
Flag: NONE pre-production
Change-Id: If2dd8861a5e06dd59292f5f469f91b9ab9f7a3b9