(Aos) Bypass PCO waiting check for Emergency Registration
IsPcoWaitingRequired() determines if the device must wait for a PCO value. This logic supports a feature related to the operator reserved PCO (FF00H) value, which dictates the IMPU selection.

Problem: The PCO waiting logic was being incorrectly applied to Emergency Registration. This feature should only be active for Normal Registration.

Solution: Modified IsPcoWaitingRequired() to add a guard clause at the beginning of the function. If IsEmergencyType() is true, the function now immediately returns IMS_FALSE, bypassing the PCO check logic for all emergency sessions.

This ensures that the feature related to the following requirement is not applied to emergency registration:

< VZ_REQ_LEQ_LTEDATA_38475 >
If the value of the operator reserved PCO container FF00H in the ACTIVATE DEFAULT EPS BEARER CONTEXT REQUEST message for the IMS PDN is 5, then the device shall use the IMSI-based SIP URI for the IMS Public User Identity.

Bug: 439461918
Flag: NONE pre-production
Test: atest and TISS test
Change-Id: Ia1de4ee9cf62a3ba47fbe8fb0f3bdbd8a3146a58