(Isil) Send call end callbacks to the Telephony when it is ready
The `TELEPHONY_LISTENING` detail is added to indicate when the Telephony framework is ready to receive callbacks.
Based on the codes and test results, we confirmed that notifying call termination when `ImsCallSessionImpl.start`(for outgoing calls) or `ImsCallSessionImpl.onIncomingcallNotified`(for incoming calls) is invoked results in normal termination.
For incoming calls, if the call is already terminated when `ImsCallSessionImpl.onIncomingcallNotified` is called, the UE notifies via callback immediately.
Although it is currently impossible for a user-initiated outgoing call to terminate at the point when ImsCallSessionImpl.start is called, that detail is applied for future use.

Bug: 453575422
Test: atest ImsStackJavaTests
Test: TISS simulation(MO, MT, call swap, merge, see the test list in the comment#15)
Flag: NONE pre-production
Change-Id: I2108d9bcdf58846cb50f34d8da2a4f186116d366