Fix: Default to full bitrate/bandwidth while receiving offer without BW/BR included
Fix of failing to negotiate the EVS codec with incoming VoLTE calls when the incoming invite has no bandwidth (BW) or no bitrate (BR) parameters in SDP offer.
Per TS 26 445, If bw/br is not present, all bandwidths/bit-rate(s) consistent with the negotiated bit-rate(s)/bandwidths are allowed in the session unless bw-send/br-send or bw-recv/br-recv is present.
Default to set to the full bitrate/bandwidth to consistent with the supported bandwidths/bit-rate(s)

Bug: 442684652
Test: atest AudioSdpParserTest
Flag: EXEMPT BUGFIX
Change-Id: I73faeee4d9655d26666c300a3de37810647a8266