(Mtc) not to set m_bLocalResourceConfirmedInitially when on the forking session
When MtcSession is created, QosInfo is also created for each session. At this time, while performing CreateStatusRecords, there is a process of setting the bLocalResourceConfirmedInitially variable to check whether local QoS information was sent out as sendrecv in the initial INVITE. This variable is set not only in the initial MtcSession, but also in the case of a forking session. it's about the orginal session not for forking session, fixed not to set m_bLocalResourceConfirmedInitially when on the forking session

test: MO and forking session in TISS.
Test: atest ImsStackNativeEnablerMtcTests
Bug: 460292061
Flag: NONE pre-production
Change-Id: I75a78b5c157214f0e2e3fcb99e5bbaf0a4842639