(Mtc) Change the condition for resetting `RegistrationThrottlingTime`
The `RegistrationThrottlingTime` was originally added(ag/34042097) under the assumption that `MtcRadioChecker::Check` would be called only once during call origination.
Due to a change in the trigger timing of StartImsTraffic (b/433869994), `MtcRadioChecker::Check` is now called twice by default during call origination.
To avoid the possibility of losing that information, we are modifying the code to reset `RegistrationThrottlingTime` once per call, as was done previously.

Test: atest ImsStackNativeEnablerMtcTests
Bug: 433869994
Flag: NONE pre-production
Change-Id: I4a0041307a0a908cbc4ae665a021f3efcde373ba