(Mtc) Fix bug : not to send precondition confirmation for not used media
This commit fixes the issue that unnecessary UPDATE requests for the QoS
precondition confirmation are sent for the media which doesn't use
precondition attribute.
When only the audio media includes QoS precondition attributes and other
media types do not, the local QoS status for those other media types is
not updated in the UPDATE message. However, the MTC then assumes it has
never sent the QoS precondition confirmation for those media types and
repeatedly sends UPDATE message without the attribute.

This fix modifies the behavior so that if the peer does not support the
QoS precondition attribute for a specific media type, the MTC will not
attemp to send a confirmation for that media.

Bug: 460732699
Test: atest ImsStackNativeEnablerMtcTests
Flag: NONE pre-production
Change-Id: I83b5d42929010adb80f2f8e2bdca5545a20c3898