(Aos) Release EPDN after E-Call if device is on E-Attach
This change introduces a new carrier configuration to control whether
the Emergency PDN (EPDN) is torn down after an emergency call when the
device is in an Emergency Attach state.

If the config `IsReleaseEPdnUponECallEndIfEAttach` is true and the
device reports being emergency-attached, the EPDN will be released. This
behavior is a carrier requirement.

- Refactored `IsReleaseEmergencyPdnUponEmergencyCallEnd` for better readability by using a single return point for the IPCAN check logic.
- Added comprehensive unit tests to cover all logical branches of the new condition.

Bug: 461352293
Flag: NONE pre-production
Test: atest and manual test
Change-Id: I39f03bc3acc9b3340c0e0191e88b74bece64d1ee