(Aos) Receive call state change asynchronously as it is now
We considered overriding IsSynchronousCallRequired to ensure that OnCallSessionReleased is invoked synchronously. However, upon review, it was confirmed that OnCallSessionReleased is already invoked synchronously regardless of the IsSynchronousCallRequired.
Therefore, we are keeping the current behavior where other call state changes are received asynchronously and will not override IsSynchronousCallRequired.

Flag: DOCS_ONLY
Bug: 408340452
Test: build
Change-Id: Iec494cfab339f24adf0976652f05f07b61fb43cc