(SmsStack) Fix NPE in ImsSmsImpl sendSms
This change fixes a critical NullPointerException in ImsSmsImpl that
occurs when sending an SMS during an IMS stack reconfiguration (LTE ->
IWLAN).

The crash may have caused by a race condition between the Binder thread
handling sendSms() and the Main thread handling unbindCallApp(). The
unbind process calls clear(), which sets the SmsTransferLayer (mSmsTL)
to null. Because sendSms() is not synchronized, it can pass the 'mReady'
check and then attempt to dereference 'mSmsTL' after it has been
nullified, resulting in a crash.

Additionally, this fix prevents a "zombie state" where onReady() could
set 'mReady = true' even if 'mSmsTL' is null (if init() was skipped or
cleared), which would also lead to this crash.

Bug: 462446792
Test: atest ImsStackJavaTests
Flag: NONE pre-production
Change-Id: I30b152ca48c52b11bc20e8bfb282302d65179871