(Aos) Update whether connected over IWLAN
AosNetTracker caches whether the connection over IWLAN is established and is updated only when connected.
This cached value is used to determine whether the network type is IWLAN when notified of deregistered status.
If connected and disconnected over IWLAN, then tried to connect via NR but failed, the network type may be determined as IWLAN.
This CL is modified to update the cached value whenever the connection status changes.

Flag: NONE pre-production
Bug: 464116381
Test: atest ImsStackNativeTests
Change-Id: I11dc86c5025321416f179e384223408ced0899a1