(Mtc) Update SUPP_TYPE_CW based on Alert-Info in 180 Response
This changes to update the SUPP_TYPE_CW supplementary service when a 180
Response is received. The update is based on the value of the Alert-Info
header, which is then passed to Telephony as the
[CW](https://source.corp.google.com/h/googleplex-android/platform/superproject/main/+/main:frameworks/opt/telephony/src/java/com/android/internal/telephony/gsm/SuppServiceNotification.java;l=82?q=CODE_1_CALL_IS_WAITING&ss=h%2Fgoogleplex-android%2Fplatform%2Fsuperproject%2Fmain%2F%2B%2Frefs%2Fheads%2Fmain&pli=1)
value.

Additionally, the change modifies the code to mock IMessageUtils instead
of mocking IMessage within MtcSupplementaryServiceTest.

Test: atest ImsStackNativeEnablerMtcTests
Test: MO call w/ and w/o 180 Alert-Info simulation
Bug: 463815652
Flag: NONE pre-production
Change-Id: Ib4e1a1d7524c5d781ab3790b8a4f6bd5453b4305