(Mtc) Handle History-Info in provisional responses
Currently, History-Info is only utilized when receiving an incoming INVITE. However, History-Info can also be included in Provisional Responses, and this information needs to be passed to the Telephony layer.

This change enhances the `UpdateCdiv()` function in `MtcSupplementaryService` to process History-Info when a Provisional Response is received, and modifies the `OutgoingState` to invoke this function.

Test: atest ImsStackNativeEnablerMtcTests
Test: MO call w/ and w/o 181 History-Info simulation
Bug: 463816399
Flag: NONE pre-production
Change-Id: Ic21660be828c413abe9673bfe34c7bca263dc1a7