(platform) Use a default MAC address in PANI header when unavailable
When the cross-SIM calling feature is enabled, if IMS registration
proceeds via IWLAN over other SIM's data network, there is an issue
where the P-Access-Network-Info(PANI) header is not set due to the
absence of a Wi-Fi MAC address.
To correct this, this change modifies the platform to recognize a
default MAC address(00:00:00:00:00:00) when the Wi-Fi MAC address is
unavailable, ensuring the PANI header is set correctly.
Additionally, this modification ensures that the PANI header is not
omitted even when the device is actually using a Wi-Fi network, but
lacks a MAC address by applying the same default MAC address in that
scenario.

Bug: 460796663
Flag: EXEMPT BUGFIX
Test: atest ImsStackNativeTests
Test: manual (verify PANI header is included in SIP REGISTER request
when cross-SIM calling is enabled)

Change-Id: Id2cda1d270d2ba2de5b81689dbc78c09bafef21d