(Mtc) Include user=phone parameter in user entity comparison
For multi-device supporting carriers like VZW, conference calls must be supported for two devices sharing the same MSISDN. In these cases, the conference focus provides two user entity values in the conference info package that differ only by the presence or absence of the 'user=phone' parameter.

To correctly distinguish between these entities, this commit modifies the comparison logic to additionally check for the presence of the 'user=phone' parameter for user entities with the same number.

- Comparison is limited to this specific case to minimize changes (other URI parameters are not compared).
- Comparing the endpoint-entity is not considered due to the large scope of modifications required.

Test: atest ImsStackNativeEnablerMtcTests
Test: anonymous user entity, same msisdn w/ and w/o user=phone
Bug: 463842563
Flag: NONE pre-production
Change-Id: I32c480eb93e852e5e42f17dccf5461744b91a676