(Mtc) Not block session update if UE receives the request with both video and text port 0 for policy TEXT_VIDEO_NOT_ALLOWED
The newtork downgrade the RTT call and append additional video
capability with port 0 in the RE-INVITE request. VZW device reject the
downgrade request due to policy TEXT_VIDEO_NOT_ALLOWED.

To cover this anomalous network behavior, we do not block the case where both video and text ports are 0 in the SDP offer.

- GetRemoteCallTypeIncludingInactiveMedia() is replaced by HasVideoRttSdp()

Bug: 459916845
Test: atest ImsStackNativeTests
Test: Verify the issue scenario reported in b/459916845#comment1 and b/467400357#comment43 with patche#12
Flag: NONE pre-production
Change-Id: I1fb277815782f36f186128025d8ebd55e019d068