(Aos) proceed with carrier's default retry mechanism after the 3GPP standard retry
The UE is required to follow 3GPP standard handling for specific SIP error codes first.
However, if the conditions for the standard 3GPP retry are not met (e.g., missing required headers),
the UE should proceed to the carrier's default retry mechanism instead of stopping or incorrectly handling the failure.

Requirement: GID-MTRREQ-482391, GID-MTRREQ-482392, GID-MTRREQ-482394
The retry config is configured for that carrier
- ims.extra_reg_err_policy_int = CarrierConfig::Ims::ERROR_POLICY_PDN_REACTIVATED
- ims.rereg_retry_err_code_for_init_reg_with_same_pcscf_int_array

Flag: NONE pre-production
Bug: 460362317
Test: test TC, manual test by using TISS and atest ImsStackNativeTests
Change-Id: Id991a3bd7dbc558845048b5df8e776d1ca17be5d