(MTC) Resolve RAT is not updated to Telephony from MTC when P-CSCF is changed after a HO between Cellular and WiFi
When the P-CSCF changes after a Handover between Cellular to Wi-Fi,
if ims.keep_existing_pcscf_on_pcscf_change_during_the_call_bool is set to false, pending reg recovery will be held during an active IMS call. In this case, MTC will not be notified of the new IPCAN category, unlike the scenario where no P-CSCF change occurs.
It will cause the RIL set Speech to wrong state to the modem and eventually lead to a call drop.

With this fix, AosApplication will notify MtcService of the new IPCAN category immediately when it receives the IPCAN change event in the CONNECTED or UPDATED states, regardless of whether the PCSCF has changed.
As a result, this allows MtcCall to update the RAT to Telephony when the IPCAN changes in this scenario.

For more details, please refer to go/mtc-resolvecalldropduetopcscfchange

Bug: 461380859

Test: atest ImsStackNativeTests
Test: Lab test and manual test PASS
Flag: NONE pre-production
Change-Id: I3bd5feaf8231be0165eed741c209f9181d112201