(Mtc) Set `TERMINATED` status with an auto rejected call
Before the fix in ag/36607524, when a call was auto-rejected, setting the state to `TERMINATED` in `alertUser` caused an issue because the Telephony framework referenced that state while processing the incoming call.
Now, `alertUser` is only called if the Telephony framework has correctly acknowledged the call as an incoming call. Therefore, the actual state, TERMINATED, is applied in the case of an auto-rejected call.

Bug: 467387101
Test: atest ImsStackNativeTests
Test: simulated the issue case with TISS
Flag: NONE pre-production
Change-Id: Id0e5c8dc1f95c34c6abdc4fd174b99a3c9854332