(config) Classify parent carriers of Nuro Mobile with MNC
The SIM properties of Nuro Mobile (canonical_id: 2432) have preferred_apn added to the MCC/MNC of Dcomo, Kddi, and Softbank.
Therefore, when Nuro Mobile SIM is used, it is necessary to follow the settings of each operator according to the MNC information.
This CL modified to set the parent carrier of Nuro Mobile according to MNC information.

Bug: 458564057
Test: build
Flag: NONE resource only update
Change-Id: I34668f084b8394e90989413cf8c333a5649b543a