(Mtc) Fix incorrect USSI body handling during network-initiated USSI
This update corrects a logic error in handling network-initiated USSI requests.
- As-Is: Attempted to use the USSI body from the ACK message.
- To-Be: Uses the USSI body from the initial INVITE message.

Since the communication path between Telephony and the IMS stack is not established at the moment the INVITE is received, the MTC first sends a 200 OK. Only after receiving the ACK does it process the USSI body.

However, because the USSI body is contained in the initial INVITE (not the ACK), the logic has been updated to retrieve the body from the INVITE message for processing after the ACK is received.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulation by TISS for network initiated USSI
Bug: 468979445
Flag: NONE pre-production
Change-Id: I4d21f87fac2a4b102dbaf7e1bb8574bd9cf80bad